+++ client/src/pages/PorHorror/PHGame.tsx
@@ -1,11 +1,10 @@
 import { useContext, useEffect, useRef } from "react";
 import { CanvasDrawer, MainScreen } from "../../services/drawer";
 import { IBasePage, PAGES } from "../PageManager";
-import useKeyboard from "./hooks/useKeyboard";
-import { Input } from "../../services/engine/structures";
 import { Button } from "../../components";
 import { ServerContext, StoreContext } from "../../App";
 import { TUpdateSceneResponse } from "../../services/server/types";
+import { Input, useKeyboard } from "../../services/input";
 
 const PHGame: React.FC<IBasePage> = (props: IBasePage) => {
     const server = useContext(ServerContext);
