+++ client/src/services/engine/structures/index.ts
@@ -1,5 +1,4 @@
-import Input from "./Input/Input";
 import Physic from "./Physic/Physic";
 import Scene from "./Scene/Scene";
 
-export { Scene, Physic, Input };
+export { Scene, Physic};
