+++ client/src/services/input/index.ts
@@ -1,0 +1,4 @@
+import Input from "./Input";
+import useKeyboard from "./hooks/useKeyboard";
+
+export { Input, useKeyboard };
